Modification:
1. Change the testbed circuit.
The TX and RX Arduinos now use "Serial1" to synchronize between
experimental traces, which is a more robust way than the previous one
(the voltage of two digital pins are not stable enough during
transmission, and will lead to unexpected save file break).

2. Change the Arduino codes according to the above modification.

3. Fix some bugs in the "plain0" TX Arduino codes.

4. Fix some bugs in the data processing and figure plotting code.
Some data processing were missing, and figure plotting used wrong
results.

5. Add a main file "main_sim_txrx.m" which calls the simulation
functions used in the early stage of the project, which is not fully
compatible with the latest decoder.

6. About documentation.
- Add detailed description on the two MATLAB toolboxes used in the code,
which are highly recommended. The modification of corresponding codes to
avoid these toolboxes are also provided.
- Fix the description about the symbol interval of the data.